pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opentrade"
version = "0.1.0"
description = "AI-Powered Cryptocurrency Trading System with Self-Evolving Strategies"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "OpenTrade Team", email = "contact@opentrade.ai"}
]
keywords = ["trading", "cryptocurrency", "ai", "trading-bot", "hyperliquid"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core
    "python-dateutil>=2.8.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # Async
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    
    # Trading
    "ccxt>=4.0.0",
    
    # AI/ML
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    
    # Optional: LangGraph
    "langgraph>=0.0.1; python_version >= '3.11'",
    
    # Optional: Vector DB
    "qdrant-client>=1.6.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    
    # Monitoring
    "prometheus-client>=0.19.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
]
all = [
    "opentrade[dev]",
    "structlog>=24.0.0",
    "ta-lib>=0.4.0",
]

[project.scripts]
opentrade = "opentrade.cli.main:app"

[project.urls]
Homepage = "https://github.com/waeaiteam/opentrade"
Repository = "https://github.com/waeaiteam/opentrade"
Issues = "https://github.com/waeaiteam/opentrade/issues"
Documentation = "https://docs.opentrade.ai"

[tool.setuptools.packages.find]
where = ["."]
include = ["opentrade*"]
exclude = ["tests*", "scripts*", "docs*", "frontend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = "\\.pyi?$"
extend-exclude = """
/(
    \\.eggs
    | \\.git
    | \\.hg
    | \\.mypy_cache
    | \\.tox
    | \\.venv
    | _build
    | buck-out
    | build
    | dist
)/
"""

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["opentrade"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[project.entry-points."opentrade.plugins"]
hyperliquid = "opentrade.plugins.exchanges:HyperliquidPlugin"
binance = "opentrade.plugins.exchanges:BinancePlugin"
